requirements.txt
streamlit
groq
pandas 
requests 
python-dotenv
